segundo commit

+++ src/app/model/Curso.java
@@ -2,8 +2,8 @@ package app.model;
 
 public class Curso {
 	private String nombre;
-	private double horas;
-	public Curso(String nombre, double horas) {
+	private int horas;
+	public Curso(String nombre, int horas) {
 		this.nombre = nombre;
 		this.horas = horas;
 	}
@@ -13,10 +13,10 @@ public class Curso {
 	public void setNombre(String nombre) {
 		this.nombre = nombre;
 	}
-	public double getHoras() {
+	public int getHoras() {
 		return horas;
 	}
-	public void setHoras(double horas) {
+	public void setHoras(int horas) {
 		this.horas = horas;
 	}
